Choose a Lattice Map slot by rounded 12-TET pitch, not by key number (#892)
Map mode read the raw key number and added its correction to whatever
arrived, so a keyboard already sending its own tuning was tuned twice: an
E arriving 13.686 cents flat onto a map whose E is 5/4 was bent another
13.686 cents (#891).

An onset's whole sounding pitch — key, per-note tuning and channel bend —
now selects the slot by rounding to the nearest 12-TET semitone, and the
map states an absolute pitch rather than an increment. The emitted
correction is the difference from what arrived to where the map puts it,
so an incoming detune is spent choosing a slot instead of accumulating.
Plain 12-TET input is unchanged, which is why every existing map test
passes untouched.

Reading a standing bend as part of the arriving pitch is the rule
Adaptive already applies, so the two engines never disagree about what
was played; `incoming` is now computed once and shared by both branches.
The cost is stated in docs/lattice-maps.md: a bend held at the attack is
absorbed into the frozen correction, and the wire carries no way to tell
a tuning event from a performed bend.

`node`/`correction` take an i64 key, since a rounded pitch is not a u8.

+++ crates/harmonigraph-ui/src/panes/lattice.rs
@@ -241,7 +241,7 @@ pub(crate) fn draw_lattice(
             && state.runtime.neighbourhood.has_context();
         for node in &scene.nodes {
             let assigned = active_map
-                .and_then(|map| (0..12u8).find(|&midi| map.node(midi) == node.lattice_pos));
+                .and_then(|map| (0..12i64).find(|&midi| map.node(midi) == node.lattice_pos));
             let outlined = assigned.is_some()
                 || adaptive && state.runtime.neighbourhood.nodes.contains(&node.lattice_pos);
             let candidate = maps.is_some_and(|m| m.editing())
@@ -279,7 +279,7 @@ pub(crate) fn draw_lattice(
         {
             if let Some(destination) = state.surfaces.hovered.filter(|_| maps.edit_shape) {
                 let midi =
-                    harmonigraph_core::lattice_map::LatticeMap::midi_class(destination) as u8;
+                    harmonigraph_core::lattice_map::LatticeMap::midi_class(destination) as i64;
                 let mut map = maps.playback.map.unwrap_or_default();
                 let old = map.correction(midi, state.runtime.tuning);
                 map.replace(destination);
